pyproject.toml
[tool.poetry]
name = "naturalboldcopyrightinfringement"
version = "0.1.0"
description = ""
authors = ["NicholasNelson <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
pyTermTk = "^0.14.0-alpha.0"
replit = "^3.2.5"
pycrypto = "^2.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
